Extract status/error UI cluster into engine/src/ui/status.ts
Moves showError, clearError, showStatus, showColorMappingStatus,
switchTab, showKeyboardShortcuts, createShortcutsUI, and
updateWelcomeMessageVisibility out of PointCloudVisualizer. Most of
these were already fully stateless DOM/console functions; only
updateWelcomeMessageVisibility needed a narrow host parameter
(spatialFiles.length, isFileLoading), and the two shortcuts-panel
functions take callbacks for the one piece of visualizer state they
touch (createShortcutsUI/updateControlStatus) instead of a full host
interface.

main.ts: 13,470 -> 13,286 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
